+++ src/main/java/interfaceadapters/generateMazeInterfaceAdapters/MazeGenerationFailed.java
@@ -1,4 +1,4 @@
-package interface_adapters;
+package interfaceadapters.generateMazeInterfaceAdapters;
 
 public class MazeGenerationFailed extends RuntimeException {
     public MazeGenerationFailed(String error) {
