+++ src/main/java/usecases/generateMazeUseCase/MazeGeneratorInputBoundary.java
