+++ src/main/java/usecases/generateMazeUseCase/MazeGeneratorOutputBoundary.java
