+++ src/main/java/usecases/generateMazeUseCase/MazeGeneratorRequestModel.java
