+++ src/main/java/usecases/generateMazeUseCase/MazeGeneratorResponseModel.java
