+++ src/main/java/usecases/generateMazeUseCase/MazeRepresentation.java
